% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This Function creates visualization of accidents as points in United States Geographical Map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State ID from the file.}

\item{year}{Year corresponding to the file.}
}
\value{
plot the accident data as points.
}
\description{
This Function creates visualization of accidents as points in United States Geographical Map.
}
\details{
If the input \code{state.num} is not present will show the error \code{"invalid STATE number: state.num"}, If no data is present in the file with respect to the input \code{state.num} and \code{year} it will print the message \code{"no accidents to plot"}.
}
