% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read .R
\name{fars_read}
\alias{fars_read}
\title{This Function reads the csv file \code{filename} from the local directory and returns as \code{tibble}}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name or path of the csv file.}
}
\value{
A tibble converted from the csv file.
}
\description{
This Function reads the csv file \code{filename} from the local directory and returns as \code{tibble}
}
\details{
If \code{filename} does not exist error will be shown as \code{Error in fars_read("filename") : file 'filename' does not exist}
}
\examples{
\dontrun{
# Read data from a CSV file and display the first few rows
data <- fars_read("path/to/fars_data.csv")
head(data)
}
}
