% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This Function reads the csv files corresponding to code{years} from the local directory and returns tibbles with two column MONTH and year for each file.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the years.}
}
\value{
A list containing tibble for each csv file.
}
\description{
This Function reads the csv files corresponding to code{years} from the local directory and returns tibbles with two column MONTH and year for each file.
}
\details{
If there are no files specific to any of the years \code{"invalid year: "} warning will be shown in the console and the corresponding list element will #' be NULL
}
