% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read .R, R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This Function creates summarization of the files corresponding to the \code{years}.}
\usage{
fars_summarize_years(years)

fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing the years.}
}
\value{
summarized table from the csv files.

summarized table from the csv files.
}
\description{
This Function creates summarization of the files corresponding to the \code{years}.

This Function creates summarization of the files corresponding to the \code{years}.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014,2015))

}
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014,2015))

}
}
