% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read .R, R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{This Function creates string.}
\usage{
make_filename(year)

make_filename(year)
}
\arguments{
\item{year}{A vector or single \code{integer} valued input, floating point values will be converted to \code{integer}.}
}
\value{
A string.

A string.
}
\description{
This Function creates string.

This Function creates string.
}
\examples{
make_filename(2023)
make_filename(c(2022,2023,2024))
make_filename(2023)
make_filename(c(2022,2023,2024))
}
