% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{This Function creates string.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A vector or single \code{integer} valued input, floating point values will be converted to \code{integer}.}
}
\value{
A string.
}
\description{
This Function creates string.
}
